backend/scent/requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
Django==3.2.7
django-extensions==3.1.3
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
ipython==7.27.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
packaging==21.0
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
Pygments==2.10.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
traitlets==5.1.0
uritemplate==3.0.1
urllib3==1.26.7
wcwidth==0.2.5
django-cors-headers==3.7.0